refactor(translation): update translation to hook
- update translation to hook, since it uses react states and is a technical hook, not util
